requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
atpublic==5.0
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
build==0.10.0
CacheControl==0.13.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==43.0.3
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
Django==4.2.16
dlib==19.24.6
dulwich==0.21.7
eval_type_backport==0.2.0
exceptiongroup==1.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0
faiss-cpu==1.9.0.post1
fastapi==0.115.2
filelock==3.16.1
filetype==1.2.0
Flask==2.3.3
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.17.3
keyring==24.3.1
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-ollama==0.2.0
langchain-openai==0.2.9
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.143
llama-cloud==0.1.5
llama-index==0.12.2
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.2
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.15
llama_cpp_python==0.3.2
lxml==5.3.0
MarkupSafe==3.0.1
marshmallow==3.23.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
ollama==0.4.2
openai==1.55.0
opencv-python==4.10.0.84
orjson==3.10.11
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pexpect==4.9.0
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pkginfo==1.11.2
platformdirs==3.11.0
plotly==5.24.1
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.6.0
propcache==0.2.0
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==2.10.0
pyOpenSSL==24.2.1
pypdf==5.1.0
PyPDF2==3.0.1
pyproject_hooks==1.2.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-multipart==0.0.12
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowflake==1.0.2
snowflake-connector-python==3.12.3
snowflake._legacy==1.0.0
snowflake.core==1.0.2
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.39.2
stripe==11.3.0
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.0.2
tomlkit==0.13.2
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.66.5
transformers==4.46.3
trove-classifiers==2024.10.21.16
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.31.1
virtualenv==20.27.0
Werkzeug==3.0.4
wrapt==1.17.0
yarl==1.16.0
zipp==3.20.2
